Silenced a warning and Fixed a bug on a parameter. Now EtoileMenus are shown again and work
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@8262 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Bundles/EtoileWildMenus/GSHorizontalMenuView.m
@@ -159,7 +159,7 @@ typedef struct _GSCellRect {
     }
 }
 
-- (NSRect) rectOfItemAtIndex: (int)index
+- (NSRect) rectOfItemAtIndex: (NSInteger)index
 {
   GSCellRect aRect;
